restructure components: move CustomSelect to its own folder and add MenuAnnouncement
